implemented the boolean and vector model.
Basically, did the pre processing of the documents and the queries. Then calculated the tf-idf marices for both documents and queries and then did the same for calculation of bm25 matrices as well. Then further using these matrices, and using cosine similarity, calculated the relevance of each document given the query.